made visual distinction between double value and single value TDF and bigwig files
git-svn-id: svn://svn.code.sf.net/p/genomeview/code/genomeview/trunk@1726 6512a8d4-c03c-0410-a19d-ba0f7ff4f5ce

+++ net/sf/genomeview/gui/viztracks/hts/PileupTrack.java
@@ -165,7 +165,7 @@ public class PileupTrack extends Track {
 						"Input the height of the new threshold line", "Input value", JOptionPane.QUESTION_MESSAGE);
 				if (in != null) {
 					try {
-						Integer d = Integer.parseInt(in);
+						Double d = Double.parseDouble(in);
 						ptm.addLine(new Line(d));
 					} catch (Exception ex) {
 						log.log(Level.WARNING, "Unparseble value for maximum in PileupTrack: " + in, ex);
